+++ Frontend/src/app/model/policy.ts
@@ -23,3 +23,11 @@ export interface PolicyReceive {
     email: string
 }
 
+export interface Policy {
+    id: number;
+    coverName: string;
+    policyType: string;
+    sumInsured: number;
+    description: string;
+  }
+
