--- Frontend/src/data/mock-policy.ts
@@ -1,33 +1,0 @@
-import { Policy } from './policy';
-
-export const Policies: Policy[] = [
-  {
-    "id": 1,
-    "policyType": "Health Insurance",
-    "coverName": "UHGHI Family Cover",
-    "description": "This plan covers insurance for your family",
-    "sumInsured": 12000
-  },
-  {
-    "id": 2,
-    "policyType": "Health Insurance",
-    "coverName": "UHGHI Individual Insurance",
-    "description": "This plan covers safety for an individual",
-    "sumInsured": 12000
-  },
-  {
-    "id": 3,
-    "policyType": "Life Insurance",
-    "coverName": "UHGCP Life Risk Cover",
-    "description": "Make your family secure, because lift isn't",
-    "sumInsured": 12000
-  },
-  {
-    "id": 4,
-    "policyType": "Term Insurance",
-    "coverName": "UHGTI Term Cover",
-    "description": "Best term insurance in the market, providing highest sum insured",
-    "sumInsured": 12000
-  }
-]
-
