--- Frontend/src/data/policy.json
@@ -1,30 +1,0 @@
-[
-  {
-    "policyId": 1,
-    "policyType": "Health Insurance",
-    "coverName": "UHGHI Family Cover",
-    "Description": "This plan covers insurance for your family",
-    "sumInsured": 12000
-  },
-  {
-    "policyId": 2,
-    "policyType": "Health Insurance",
-    "coverName": "UHGHI Individual Insurance",
-    "Description": "This plan covers safety for an individual",
-    "sumInsured": 12000
-  },
-  {
-    "policyId": 3,
-    "policyType": "Life Insurance",
-    "coverName": "UHGCP Life Risk Cover",
-    "Description": "Make your family secure, because lift isn't",
-    "sumInsured": 12000
-  },
-  {
-    "policyId": 4,
-    "policyType": "Term Insurance",
-    "coverName": "UHGTI Term Cover",
-    "Description": "Best term insurance in the market, providing highest sum insured",
-    "sumInsured": 12000
-  }
-]
