--- Frontend/src/data/policy.ts
@@ -1,7 +1,0 @@
-export interface Policy {
-  id: number;
-  coverName: string;
-  policyType: string;
-  sumInsured: number;
-  description: string;
-}
